Refactor out LayerList component out of LayersEditor
